final-pjt/final_pjt_back/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
dj-rest-auth==6.0.0
Django==4.2.6
django-allauth==0.63.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.1
exceptiongroup==1.2.0
executing==2.0.1
idna==3.7
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.12
